Refactor, las cajas se van al layout
